Created hidden labels function, moved extension function
