+++ src/index.html
@@ -500,6 +500,7 @@
 						data-makeimageslinks="true"
 						data-altformats="Alternate Formats"
 						data-altlanguages="Alternate Languages"
+						data-hiddenlabels="data-altlanguages,data-imagefield,topics"
 						data-imageposition="left">
 				</sort-filter-root>
